+++ Sources/PaymentButtons/PaymentButton.swift
@@ -112,31 +112,22 @@ public class PaymentButton: UIButton {
     public private(set) var label: PaymentButtonLabel?
 
     private var imageHeight: CGFloat {
-        // For pay later or paypal credit return different image height
         switch size {
-        case .mini:
-            return 20.0
 
-        case .miniWithWordmark:
+        case .mini:
             return 24.0
 
-        case .collapsed:
-            return 15.0
-
-        case .expanded:
+        case .standard:
             return 20.0
-
-        case .full:
-            return 26.0
         }
     }
 
     private var supportsPrefixLabel: Bool {
         switch size {
-        case .mini, .miniWithWordmark, .collapsed:
+        case .mini:
             return false
 
-        case .expanded, .full:
+        case .standard:
             if let label = label {
                 return label.position == .prefix
             }
@@ -145,15 +136,11 @@ public class PaymentButton: UIButton {
     }
 
     private var supportsSuffixLabel: Bool {
-
         switch size {
-        case .mini, .miniWithWordmark:
+        case .mini:
             return false
 
-        case .collapsed:
-            return fundingSource == .payLater
-
-        case .expanded, .full:
+        case .standard:
             if let label = label {
                 return label.position == .suffix
             }
@@ -251,9 +238,8 @@ public class PaymentButton: UIButton {
     // MARK: - Override Function
     override public func layoutSubviews() {
         super.layoutSubviews()
-        if size == .mini || size == .miniWithWordmark {
-            let minValue = min(containerView.bounds.width, containerView.bounds.height)
-            containerView.layer.cornerRadius = minValue / 2
+        if size == .mini {
+            containerView.layer.cornerRadius = 4.0
         } else {
             containerView.layer.cornerRadius = edges.cornerRadius(for: containerView)
         }
@@ -267,7 +253,7 @@ public class PaymentButton: UIButton {
 
         // Size to fit
         switch size {
-        case .mini, .miniWithWordmark:
+        case .mini:
             let maxValue = max(image.size.width, image.size.height)
             imageView.bounds = CGRect(
                 x: 0,